rename packages 'ru.skypro.homework' to 'com.evgeniyfedorchenko.simpleavito'
