+++ src/main/java/com/evgeniyfedorchenko/simpleavito/SimpleAvitoApplication.java
@@ -1,0 +1,13 @@
+package com.evgeniyfedorchenko.simpleavito;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SimpleAvitoApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(SimpleAvitoApplication.class, args);
+  }
+
+}
