+++ src/main/java/com/evgeniyfedorchenko/simpleavito/dto/Ad.java
@@ -1,0 +1,13 @@
+package com.evgeniyfedorchenko.simpleavito.dto;
+
+import lombok.Data;
+
+@Data
+public class Ad {
+
+    private int author;
+    private String image;
+    private int pk;
+    private int price;
+    private String title;
+}
