+++ src/main/java/com/evgeniyfedorchenko/simpleavito/dto/Ads.java
@@ -1,0 +1,13 @@
+package com.evgeniyfedorchenko.simpleavito.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class Ads {
+
+    private int count;
+
+    private List<Ad> results;
+}
