+++ src/main/java/com/evgeniyfedorchenko/simpleavito/dto/Comment.java
@@ -1,0 +1,15 @@
+package com.evgeniyfedorchenko.simpleavito.dto;
+
+import lombok.Data;
+
+@Data
+public class Comment {
+
+    private int author;
+    private String authorImage;
+    private String authorFirstName;
+    private long createdAt;
+    private int pk;
+    private String text;
+    
+}
