+++ src/main/java/com/evgeniyfedorchenko/simpleavito/dto/Login.java
@@ -1,0 +1,11 @@
+package com.evgeniyfedorchenko.simpleavito.dto;
+
+import lombok.Data;
+
+@Data
+public class Login {
+
+    private String username;
+    private String password;
+
+}
