+++ src/main/java/com/evgeniyfedorchenko/simpleavito/dto/NewPassword.java
@@ -1,0 +1,15 @@
+package com.evgeniyfedorchenko.simpleavito.dto;
+
+import jakarta.validation.constraints.Size;
+import lombok.Data;
+
+@Data
+public class NewPassword {
+
+    @Size(min = 8, max = 16)
+    private String currentPassword;
+
+    @Size(min = 8, max = 16)
+    private String newPassword;
+
+}
