+++ src/main/java/com/evgeniyfedorchenko/simpleavito/dto/Register.java
@@ -1,0 +1,15 @@
+package com.evgeniyfedorchenko.simpleavito.dto;
+
+import lombok.Data;
+
+@Data
+public class Register {
+
+    private String username;
+    private String password;
+    private String firstName;
+    private String lastName;
+    private String phone;
+    private Role role;
+
+}
