+++ src/main/java/com/evgeniyfedorchenko/simpleavito/dto/Role.java
@@ -1,0 +1,7 @@
+package com.evgeniyfedorchenko.simpleavito.dto;
+
+public enum Role {
+
+    USER, ADMIN
+
+}
