+++ src/main/java/com/evgeniyfedorchenko/simpleavito/dto/User.java
@@ -1,0 +1,16 @@
+package com.evgeniyfedorchenko.simpleavito.dto;
+
+import lombok.Data;
+
+@Data
+public class User {
+
+    private int id;
+    private String email;
+    private String firstName;
+    private String lastName;
+    private String phone;
+    private Role role;
+    private String image;
+
+}
