+++ src/main/java/com/evgeniyfedorchenko/simpleavito/service/AuthService.java
@@ -1,0 +1,9 @@
+package com.evgeniyfedorchenko.simpleavito.service;
+
+import com.evgeniyfedorchenko.simpleavito.dto.Register;
+
+public interface AuthService {
+    boolean login(String userName, String password);
+
+    boolean register(Register register);
+}
