+++ src/main/java/com/evgeniyfedorchenko/simpleavito/service/AuthServiceImpl.java
@@ -1,0 +1,42 @@
+package com.evgeniyfedorchenko.simpleavito.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.provisioning.UserDetailsManager;
+import org.springframework.stereotype.Service;
+import com.evgeniyfedorchenko.simpleavito.dto.Register;
+
+@Service
+@RequiredArgsConstructor
+public class AuthServiceImpl implements AuthService {
+
+    private final UserDetailsManager manager;
+    private final PasswordEncoder encoder;
+
+    @Override
+    public boolean login(String userName, String password) {
+        if (!manager.userExists(userName)) {
+            return false;
+        }
+        UserDetails userDetails = manager.loadUserByUsername(userName);
+        return encoder.matches(password, userDetails.getPassword());
+    }
+
+    @Override
+    public boolean register(Register register) {
+        if (manager.userExists(register.getUsername())) {
+            return false;
+        }
+        manager.createUser(
+                User.builder()
+                        .passwordEncoder(this.encoder::encode)
+                        .password(register.getPassword())
+                        .username(register.getUsername())
+                        .roles(register.getRole().name())
+                        .build());
+        return true;
+    }
+
+}
