+++ src/main/java/com/evgeniyfedorchenko/simpleavito/service/CommentServiceImpl.java
@@ -1,0 +1,29 @@
+package com.evgeniyfedorchenko.simpleavito.service;
+
+import com.evgeniyfedorchenko.simpleavito.dto.Comments;
+import org.springframework.stereotype.Service;
+import com.evgeniyfedorchenko.simpleavito.dto.Comment;
+import com.evgeniyfedorchenko.simpleavito.dto.CreateOrUpdateComment;
+
+@Service
+public class CommentServiceImpl implements CommentService {
+
+    @Override
+    public Comments getComments(int id) {
+        return new Comments();
+    }
+
+    @Override
+    public Comment addComment(int id, CreateOrUpdateComment createOrUpdateComment) {
+        return new Comment();
+    }
+
+    @Override
+    public void deleteComment(int adId, int commentId) {
+    }
+
+    @Override
+    public Comment updateComment(int adId, int commentId, CreateOrUpdateComment comment) {
+        return new Comment();
+    }
+}
