--- src/main/java/ru/skypro/homework/SimpleAvitoApplication.java
@@ -1,12 +1,0 @@
-package ru.skypro.homework;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SimpleAvitoApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(SimpleAvitoApplication.class, args);
-  }
-}
