--- src/main/java/ru/skypro/homework/dto/Ad.java
@@ -1,14 +1,0 @@
-package ru.skypro.homework.dto;
-
-
-import lombok.Data;
-
-@Data
-public class Ad {
-
-    private int author;
-    private String image;
-    private int pk;
-    private int price;
-    private String title;
-}
