--- src/main/java/ru/skypro/homework/dto/Ads.java
@@ -1,13 +1,0 @@
-package ru.skypro.homework.dto;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class Ads {
-
-    private int count;
-
-    private List<Ad> results;
-}
