--- src/main/java/ru/skypro/homework/dto/Comments.java
@@ -1,12 +1,0 @@
-package ru.skypro.homework.dto;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class Comments {
-
-    private int count;
-    private List<Comment> results;
-}
