--- src/main/java/ru/skypro/homework/dto/CreateOrUpdateComment.java
@@ -1,13 +1,0 @@
-package ru.skypro.homework.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.Data;
-
-@Data
-public class CreateOrUpdateComment {
-
-    @NotBlank
-    @Size(min = 8, max = 64)
-    private String text;
-}
