--- src/main/java/ru/skypro/homework/dto/ExtendedAd.java
@@ -1,17 +1,0 @@
-package ru.skypro.homework.dto;
-
-import lombok.Data;
-
-@Data
-public class ExtendedAd {
-
-    private int pk;
-    private String authorFirstName;
-    private String authorLastName;
-    private String description;
-    private String email;
-    private String image;
-    private String phone;
-    private int price;
-    private String title;
-}
