--- src/main/java/ru/skypro/homework/dto/Login.java
@@ -1,10 +1,0 @@
-package ru.skypro.homework.dto;
-
-import lombok.Data;
-
-@Data
-public class Login {
-
-    private String username;
-    private String password;
-}
