--- src/main/java/ru/skypro/homework/dto/Role.java
@@ -1,5 +1,0 @@
-package ru.skypro.homework.dto;
-
-public enum Role {
-    USER, ADMIN
-}
