--- src/main/java/ru/skypro/homework/filter/BasicAuthCorsFilter.java
@@ -1,25 +1,0 @@
-package ru.skypro.homework.filter;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.lang.NonNull;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@Component
-public class BasicAuthCorsFilter extends OncePerRequestFilter {
-
-
-    @Override
-    protected void doFilterInternal(@NonNull HttpServletRequest request,
-                                    @NonNull HttpServletResponse response,
-                                    @NonNull FilterChain filterChain) throws ServletException, IOException {
-
-        response.addHeader("Access-Control-Allow-Credentials", "true");
-        filterChain.doFilter(request, response);
-    }
-}
