--- src/main/java/ru/skypro/homework/service/AuthService.java
@@ -1,9 +1,0 @@
-package ru.skypro.homework.service;
-
-import ru.skypro.homework.dto.Register;
-
-public interface AuthService {
-    boolean login(String userName, String password);
-
-    boolean register(Register register);
-}
